+++ Cargo.toml
@@ -7,7 +7,14 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [workspace]
-members = ["crates/*", "fn_macros"]
+members = [
+  "crates/text-buffer",
+  "crates/rune-macros"
+]
+
+[workspace.dependencies]
+quote = "1.0.26"
+syn = { version = "2.0.15", features = ["full"]}
 
 [workspace.package]
 edition = "2021"
@@ -19,7 +26,6 @@ bytecount = "0.6.3"
 druid = "0.8.3"
 fancy-regex = "0.12.0"
 float-cmp = "0.9.0"
-fn_macros = { version = "0.1.0", path = "fn_macros" }
 hostname = "0.3.1"
 memoffset = "0.9.0"
 num_enum = "0.7.1"
@@ -27,7 +33,6 @@ paste = "1.0.12"
 rand = "0.8.5"
 sptr = "0.3.2"
 streaming-iterator = "0.1.9"
-text-buffer = { version = "0.1.0", path = "crates/text-buffer" }
 titlecase = "2.2.1"
 fallible-iterator = "0.3.0"
 fallible-streaming-iterator = "0.1.9"
@@ -35,6 +40,9 @@ indexmap = "2.1.0"
 fxhash = "0.2.1"
 [target.'cfg(not(target_env = "msvc"))'.dependencies]
 tikv-jemallocator = "0.5"
+
+text-buffer = { version = "0.1.0", path = "crates/text-buffer" }
+rune-macros = { version = "0.1.0", path = "crates/rune-macros" }
 
 # [dev-dependencies]
 # backtrace-on-stack-overflow = "0.3.0"
@@ -51,8 +59,8 @@ default = []
 debug_bytecode = []
 
 [build-dependencies]
-syn = "2.0.15"
-quote = "1.0.26"
+quote = { workspace = true}
+syn = { workspace = true }
 proc-macro2 = "1.0.60"
 
 [workspace.lints.rust]
