+++ src/keymap.rs
@@ -1,5 +1,5 @@
 use crate::core::object::GcObj;
-use fn_macros::defun;
+use rune_macros::defun;
 
 // TODO: implement keymaps
 #[defun]
